Preeliminar implementation of RMSF
Wrote a draft version of RMSF. Due to the "local" nature of goMD I had
to cheat the program in order to get the "non-local" result, so it is
quite inefficient, especially regarding disk use.

Also it is not tested for bugs. I am not even sure that I am
calculating the RMSF properly, as I couldn't find a reference.
What I wrote is the square root of the norm of each cartesian point's
variance.