We now tolerate lack of mass, plus new task.
Before, in the Shape-task, if the mass of an atom
was not found, we killed the program. Now we just set
all masses to 1.0 in that case, and continue, unless
you ask the program to be strict.

There is also a new task "PlaneAngle" which, for each
frame, will obtain the angle (degrees) between the best
planes passing through 2 sets of atoms, respectively.
PlaneAngles has _not_ been tested as of this commit.