UI improvements and PlanesAngle fix
You can now use fixGMX or FixGMX as a task,
which allows you not to give any additional
parameter to gomd.

The angle in PlanesAngle is now only defined
between 0 and 90 degrees. This is because the
direction of the vector normal to a given plane
is difficult to define in our case,, and might
flip with a small change in the position of
the atoms whose moment of inertia tensor
define the plane.